Fix issues when drawing rectangles
There are two problems related to how the library draws rectangles:

The first one is related to DrawRectangleRoundedLines(), which is
different from all the other RayLib DrawRectangle*() functions because
it draws the outline outside of the region defined by the Rectangle
argument. To fix that, shrink the rectangle by the thickness plus one,
to accommodate some imprecision from RayLib.

The second issue is related to the roundness value that needs to be
calculated by the smallest dimension between width and height and
clamped between 0 and 1. With that, RAYLIB_NUKLEAR_ROUNDING_SCALE can
be dropped.